Add subscribed tags (#4)
This commit begins working on subscribed tags. It adds a new database table
subscribed_hashtags and its corresponding model, mapper, and tests.

In addition, the BotDriver is updated to remove all hardcoded references
to #scammer in favor of the new subscribed_hashtags model.

This does not flesh out the new responses as much as might be necessary.

* responses.txt - update to reflect current replacements
* USLBotDriver.java - use SubscribedHashTags instead of #scammer
* USLDatabase.java - update for SubscribedHashtagMapping
* MappingDatabase.java - add getSubscribedHashtagMapping
+ SubscribedHashtagMapping.java - first commit
+ MysqlSubscribedHashtagMapping.java - first commit
* BanHistory.java - remove references to scammer
+ SubscribedHashtag.java - first commit
* BanHistoryMappingTest.java - remove references to scammer
+ SubscribedHashtagMappingTest.java - first commit
* MysqlDatabaseTests.java - add MysqlSubscribedHashtagMappingTest
+ MysqlSubscribedHashtagMappingTest.java - first commit
* MysqlTestUtils.java - minor documentation improvements