Add regular FTP backup support
Turns out the ftp backup server uses ip-matching as its security protocol.
So this update adds regular FTP support in addition to SFTP support (in
case it ever gets added).

In addition, passwords are sanitized from logger output and the generating
backup db command is updated to fix weird behavior regarding pipe and
runtime exec on ubuntu.

Lastly, timestamps are changed from 0-time to 1970-01-01 00:00:01, which is
the earliest valid timestamp. Tests were working because my windows mysql
allows 0-dates, but running the tests on the VPS failed because it does
not.

* README.md - Add secure option to ftpbackups
* USLDatabaseBackupManager.java - handle secure option
* USLFileConfiguration.java - require secure option on ftpbackups
* MysqlBanHistoryMapping.java and
* MysqlPersonMapping.java and
* MysqlResponseMapping.java - avoid 0-dates