Refactor BanHistory (#4) and Driver (#6)
USLBotDriver now can throttle the amount of resources going to any one
subreddit. Scanning subreddits is still handled in the driver (though
differently), but propagating to subreddits is delegated to the new
class USLBanHistoryPropagator.

In addition, many model mappers were given fetchByID due to necessity
when creating the USLBanHistoryPropagator.

This is almost initial release ready, however it will still spend too much
time handling already banned users if a subreddit decides to change its
hashtags. A local history of banned users is required (as suggested in #4)