Finish resolving #6
I realized I can use the BanHistory as a fast-way to check if a user is
banned already. Not sure why I thought I couldn't earlier.

This solution will potentially not help as much as it could if the
propagator ever pms a subreddit it doesn't ban a user on. But for now,
this fixes the issue.

* USLBotDriver.java - add fast check if banned, suppress pms to subreddits
that we suppressed a ban on ( see changes to handlePropagateResult)

add fetchBanHistoryByPersonAndSubreddit to BanHistoryMapping, then add
requisite changes to its mysql mapping and its test.