Improve logging + dont propagate aggressively
Before I would propagate when some subreddits where still searching
forward, which isn't viable because we use timestamps to keep track of
what we've already looked at, and if we're adding things in the past
then that isn't going to work.

Now it will only propagate up to the point where it has information on
all monitored subreddits. There is also a text file describing what needs
to be done to add new subreddits (reset_notes.txt)